+++ agent/agent.go
@@ -2,13 +2,7 @@ package agent
 
 import (
 	"github.com/akaspin/soil/agent/allocation"
-	"github.com/akaspin/soil/manifest"
 )
-
-type Scheduler interface {
-	// Sync specific namespace with given manifests
-	Sync(namespace string, pods []*manifest.Pod) (err error)
-}
 
 type EvaluationReporter interface {
 	Sync(pods []*allocation.Pod)