+++ agent/api-v1/registry_unit_test.go
@@ -35,7 +35,7 @@ func Test_Unit_RegistryPut_Process(t *testing.T) {
 	assert.NoError(t, err)
 	defer r.Close()
 
-	var pods manifest.Pods
+	var pods manifest.Registry
 	err = (&pods).Unmarshal("public", r)
 	assert.NoError(t, err)
 
@@ -48,11 +48,11 @@ func Test_Unit_RegistryPut_Process(t *testing.T) {
 	resp, err := (&http.Client{}).Do(req)
 	assert.NoError(t, err)
 
-	var marks api_v1_types.RegistrySubmitResponse
+	var marks api_v1_types.RegistryPutResponse
 	err = json.NewDecoder(resp.Body).Decode(&marks)
 
-	assert.Equal(t, marks, api_v1_types.RegistrySubmitResponse{
-		"first":  0x515d988d1de74877,
-		"second": 0x20a35ccef17a69c5,
+	assert.Equal(t, marks, api_v1_types.RegistryPutResponse{
+		"api-test-0":0x937fb0d3a39c7353,
+		"api-test-1":0x77cf87a45dfb307d,
 	})
 }