+++ agent/public/registry.go
@@ -1,4 +1,0 @@
-package public
-
-type RegistryWatcher struct {
-}