+++ agent/registry/registry.go
@@ -1,1 +1,8 @@
 package registry
+
+import "github.com/akaspin/soil/manifest"
+
+type Consumer interface {
+
+	ConsumeRegistry(namespace string, payload manifest.Registry)
+}