+++ api/api-v1-types/registry.go
@@ -2,7 +2,8 @@ package api_v1_types
 
 import "github.com/akaspin/soil/manifest"
 
-type RegistrySubmitRequest manifest.Pods
-type RegistrySubmitResponse map[string]uint64
+type RegistryGetResponse map[string]manifest.Registry
 
-type RegistryGetResponse manifest.Pods
+type RegistryPutRequest manifest.Registry
+type RegistryPutResponse map[string]uint64
+