+++ manifest/pod_test.go
@@ -6,27 +6,16 @@ import (
 	"encoding/json"
 	"github.com/akaspin/soil/manifest"
 	"github.com/stretchr/testify/assert"
-	"os"
 	"testing"
 )
 
-func TestPods_Unmarshal(t *testing.T) {
-	var pods manifest.Pods
-	r, err := os.Open("testdata/example-multi.hcl")
-	assert.NoError(t, err)
-	defer r.Close()
-
-	err = (&pods).Unmarshal("private", r)
-	assert.NoError(t, err)
-	assert.Len(t, pods, 2)
-}
-
 func TestManifest(t *testing.T) {
-	res, err := manifest.ParseFromFiles("private", "testdata/example-multi.hcl")
+	var res manifest.Registry
+	err := res.UnmarshalFiles("private", "testdata/example-multi.hcl")
 	assert.NoError(t, err)
 
 	t.Run("parse", func(t *testing.T) {
-		assert.Equal(t, []*manifest.Pod{
+		assert.Equal(t, manifest.Registry{
 			{
 				Namespace: "private",
 				Name:      "first",
@@ -128,15 +117,11 @@ func TestManifest(t *testing.T) {
 	})
 }
 
-func TestParseFromFiles(t *testing.T) {
-	pods, err := manifest.ParseFromFiles("private", "testdata/files_1.hcl", "testdata/files_2.hcl")
-	assert.NoError(t, err)
-	assert.Len(t, pods, 3)
-}
-
 func TestManifest_JSON(t *testing.T) {
-	pods, err := manifest.ParseFromFiles("private", "testdata/json.hcl")
+	var pods manifest.Registry
+	err := pods.UnmarshalFiles("private", "testdata/json.hcl")
 	assert.NoError(t, err)
+
 	data, err := json.Marshal(pods[0])
 	assert.Equal(t, "{\"Namespace\":\"private\",\"Name\":\"first\",\"Runtime\":true,\"Target\":\"multi-user.target\",\"Constraint\":{\"${meta.one}\":\"one\",\"${meta.two}\":\"two\"},\"Units\":[{\"Create\":\"start\",\"Update\":\"\",\"Destroy\":\"stop\",\"Permanent\":true,\"Name\":\"first-1.service\",\"Source\":\"[Service]\\n# ${meta.consul}\\nExecStart=/usr/bin/sleep inf\\nExecStopPost=/usr/bin/systemctl stop first-2.service\\n\"},{\"Create\":\"\",\"Update\":\"start\",\"Destroy\":\"\",\"Permanent\":false,\"Name\":\"first-2.service\",\"Source\":\"[Service]\\n# ${NONEXISTENT}\\nExecStart=/usr/bin/sleep inf\\n\"}],\"Blobs\":[{\"Name\":\"/etc/vpn/users/env\",\"Permissions\":420,\"Leave\":false,\"Source\":\"My file\\n\"}]}", string(data))