+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## 0.3.1
 
+### API
+
+* GET /v1/registry
+* PUT /v1/registry
+
+### Clustering
+
+* Public namespace registry
+
 ## 0.3.0 (23.09.2017)
 
 ### API